수정2

+++ KOR_Polimeter/src/main/java/Webprogramming/KOR_Polimeter/web/api/repository/UserVoteRepository.java
@@ -16,5 +16,5 @@ public interface UserVoteRepository extends JpaRepository<UserVote, Integer> {
     @Query("SELECT uv FROM UserVote uv WHERE uv.user.id = :userId ORDER BY uv.voteDate DESC limit 1")
     Optional<UserVote> findLastVoteByUser(@Param("userId") long userId);
     Optional<UserVote> findByUserAndPolitician(User user, Politician politician);
-    List<UserVote> findByUserId(int userId);
+    List<UserVote> findByUserId(long userId);
 }
